+++ blender/extension/ui/animate_trajectory_operator.py
